Plug new lexing into VS
